requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
boto3==1.14.2
botocore==1.17.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
Django==3.1
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-graphql-auth==0.3.11
django-graphql-jwt==0.3.0
django-graphql-playground==1.0
django-storages==1.9.1
django-webpush==0.3.3
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
docutils==0.15
graphene==2.1.8
graphene-django==2.12.1
graphene-django-cud==0.6.4
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
gTTS==2.2.2
gunicorn==20.0.4
http-ece==1.1.0
idna==2.10
jmespath==0.10.0
ndg-httpsclient==0.5.1
Pillow==7.2.0
promise==2.3
psycopg2-binary==2.8.5
pusher==3.0.0
pusher-push-notifications==2.0.0
py-vapid==1.7.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
pywebpush==1.9.4
requests==2.24.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
whitenoise==5.2.0
